cleanup(scap,sinsp): return event flags directly
The event flags are (currently) specific to the savefile engine
but accessing them requires an awkward reacharound through
a dedicated vtable method to a field stored on each call to next().

We can simplify this a bit and express the intent better by simply
returning the flags from next() via an out pointer.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>